requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
drf-spectacular==0.27.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
